+++ package.json
@@ -1,31 +1,15 @@
 {
-	"private": true,
-	"name": "foundryvtt-devMode",
-	"version": "1.3.0",
-	"description": "",
-	"scripts": {
-		"publish": "gulp publish",
-		"lint": "prettier --write .",
-		"build": "gulp build",
-		"build:watch": "gulp watch",
-		"clean": "gulp clean"
-	},
-	"author": "",
-	"license": "",
-	"devDependencies": {
-		"@league-of-foundry-developers/foundry-vtt-types": "^0.7.9-6",
-		"@typhonjs-fvtt/eslint-config-foundry.js": "^0.8.0",
-		"archiver": "^5.1.0",
-		"chalk": "^4.1.0",
-		"fs-extra": "^9.0.1",
-		"gulp": "^4.0.2",
-		"gulp-git": "^2.10.1",
-		"gulp-less": "^4.0.1",
-		"gulp-sass": "^4.1.0",
-		"gulp-typescript": "^6.0.0-alpha.1",
-		"json-stringify-pretty-compact": "^2.0.0",
-		"sass": "^1.30.0",
-		"typescript": "^4.1.4",
-		"yargs": "^16.2.0"
-	}
+  "private": true,
+  "name": "foundryvtt-devMode",
+  "version": "1.3.0",
+  "description": "",
+  "scripts": {
+    "publish": "gulp publish",
+    "lint": "prettier --write .",
+    "build": "gulp build",
+    "build:watch": "gulp watch",
+    "clean": "gulp clean"
+  },
+  "author": "",
+  "license": ""
 }
