--- src/global.d.ts
@@ -1,5 +1,0 @@
-declare namespace Game {
-  interface Module {
-    api?: any;
-  }
-}
