+++ src/module/settings.js
@@ -2,7 +2,6 @@ import { DevModeConfig } from './classes/DevModeConfig';
 import { MODULE_ABBREV, MODULE_ID, MySettings } from './constants';
 import { setDebugOverrides } from './helpers';
 
-//@ts-ignore
 const debouncedReload = foundry.utils.debounce(() => {
   window.location.reload();
 }, 100);
