--- tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "lib": ["DOM", "ES6", "ES2020"],
-    "types": ["@league-of-foundry-developers/foundry-vtt-types"],
-    "moduleResolution": "node"
-  }
-}
